Extract out definitions from RamCacheCLFUS so we can test compression
